requirement.txt
python-dotenv
mysql-connector-python
pymongo
redis
kafka-python
pyspark